removed the ARM platform from the solution Added the common_state_machine header Finalized the Prop object and tests for it. Started work in the Hand state machine.

+++ JuggleLib/Trajectory.cpp
@@ -3,7 +3,7 @@
 #include <cmath>
 
 
-const float Trajectory::GRAVITY = 980.665;
+const double Trajectory::GRAVITY = 980.665;
